Split test_hook.py into focused test modules by feature
Replace monolithic 1549-line test_hook.py with five focused modules:
- test_git.py (389 lines): git plumbing tests
- test_driver.py (350 lines): build-driver discovery tests
- test_execution.py (293 lines): build execution + category tests
- test_pipeline.py (404 lines): pipeline + formatting tests
- test_cli.py (238 lines): CLI entry point + env parsing tests

Add TC002 to per-file-ignores for test files (pytest is runtime,
not type-only). All 107 tests pass across split files.
